MAT-6839 fixed the stickey footer styling
