+++ src/components/editMeasure/editor/CqlEditor.scss
@@ -79,7 +79,7 @@
 }
 
 .bottom-row {
-  height: 12px;
+  height: 112px;
   z-index: 35; //one above allotment?
   background: #fff;
   display: flex;
